Remaining logic for DeleteSiteController
